pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyrl"
version = "0.0.1"
authors = [
  {name="Filipo STUDZINSKI PEROTTO", email="filipo.perotto@onera.fr"},
  {name="Aymane OUAHBI", email="aymane.ouahbi@onera.fr"},
  {name="Melvine NARGEOT", email="melvine.nargeot@onera.fr"},
]
description = "Safe and Survival Reinforcement Learning"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
	'numpy',
	'scipy',
	'matplotlib',
	'pandas',
	'tqdm',
	'ray',
	'SMPyBandits',
	'gym',
	'gymnasium',
	'pygame',
	'pymdptoolbox'
]

[project.optional-dependencies]
    none = []
	basic = []
    simple = []
	tested = ['swig', 'gymnasium[box2d]', 'gymnasium[atari,accept-rom-license]', 'gymnasium[all]', 'rich', 'torch', 'stable-baselines3']
    gym_all = ['swig', 'gymnasium[box2d]', 'gymnasium[atari,accept-rom-license]', 'gymnasium[all]']
	tensorflow = ['tensorflow-cpu']
	pytorch = ['torch', 'torchvideo', 'torchaudio']
	deep_cpu = ['tensorflow-cpu', 'torch', 'torchvideo', 'torchaudio']
	tensorforce = ['tensorflow-cpu', 'tensorforce']
	sb3 = ['rich', 'torch', 'torchvideo', 'torchaudio', 'stable-baselines3']
	atari = ['gymnasium[atari,accept-rom-license]']
	all = ['tensorflow-cpu', 'tensorforce', 'torch', 'torchvideo', 'torchaudio', 'stable-baselines3', 'swig', 'gymnasium[all]']

[project.urls]
"Homepage" = "https://github.com/fsperotto/pyrl"
"Bug Tracker" = "https://github.com/fsperotto/pyrl/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["pyrl", "pyrl.mab", "pyrl.mdp", "pyrl.replay_buffer", "pyrl.environments", "pyrl.agents", "pyrl.agents.survival"]

#[tool.setuptools.packages.find]
#where = ["src"]